Added a block or attack interface to make monsterfights more fair and not seem like a dice roll (which it is any ways)

+++ TextAdventure/entities/player.java
@@ -5,7 +5,7 @@ public class player extends entity {
     public static ArrayList<String> traits = new ArrayList<String>();
     public static int roomsVisited = 0;
     public static int monstersKilled = 0;
-    public static int hp = 3; // Hit Points (deduce one at a time)
+    public static double hp = 3; // Hit Points (deduce one at a time if not blocking, half if blocking)
     public static int df = 0; // Defense (chance - 0.05 per point in Normal, 0.025 in '80s)
     public static double chanceToHit = 0.6;
